% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{Light detection and ranging}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 221 rows and 2 columns.
}
\source{
Sigrist, M. (Ed.) (1994). \emph{Air Monitoring by Spectroscopic Techniques} (Chemical Analysis Series, vol. 197). New York: Wiley.

Originally downloaded from \url{http://matt-wand.utsacademics.info/webspr/lidar.txt}. Also available in the \href{https://cran.r-project.org/web/packages/HRW/}{HRW} package.
}
\usage{
lidar
}
\description{
This data frame contains the following columns:

\code{range} distance traveled before the light is reflected back to its source.

\code{logratio} logarithm of the ratio of received light from two laser sources.
}
\details{
The \code{lidar} data frame has 221 pairs from a LIght Detection And Ranging (LIDAR) experiment.
}
\examples{
data(lidar)
with(lidar, plot(range,logratio))
}
\keyword{datasets}
