% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{mobility}
\alias{mobility}
\title{Economic mobility dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 741 rows and 43 columns.
}
\source{
This data set was based on the paper "Where is the Land of Opportunity?
The Geography of Intergenerational Mobility in the United States" (Raj Chetty,
Nathaniel Hendren, Patrick Kline, Emmanuel Saez, QJE 2014), and its replication
files. Both the paper and the data sets are available from \url{https://opportunityinsights.org.}
}
\usage{
mobility
}
\description{
Data for discussing the probability of improving one's economic situation
across 741 Metropolitan Statistical Areas in the US.
}
\details{
The variables are:\tabular{ll}{
   Variable \tab Description \cr
   Mobility \tab The probability that a child born in 1980–1982 into the lowest quintile (20\%) of household income will be in the top quintile at age 30. Individuals are assigned to the community they grew up in, not the one they were in as adults. \cr
   Population \tab in 2000. \cr
   Urban \tab Is the community primarily urban or rural? \cr
   Black \tab percentage of individuals who marked black (and nothing else) on census forms. \cr
   Racial Segregation \tab a measure of residential segregation by race. \cr
   Income Segregation \tab Similarly but for income. \cr
   Segregation of poverty \tab Specifically a measure of residential segregation for those in the bottom quarter of the national income distribution. \cr
   Segregation of affluence \tab Residential segregation for those in the top quarter. \cr
   Commute \tab Fraction of workers with a commute of less than 15 minutes. \cr
   Mean income \tab Average income per capita in 2000. \cr
   Gini \tab A measure of income inequality, which would be 0 if all incomes were perfectly equal, and tends towards 100 as all the income is concentrated among the richest individuals (see Wikipedia, s.v. "Gini coefficient"). \cr
   Share 1\% \tab Share of the total income of a community going to its richest 1\%. \cr
   Gini bottom 99\% \tab Gini coefficient among the lower 99\% of that community. \cr
   Fraction middle class \tab Fraction of parents whose income is between the national 25th and 75th percentiles. \cr
   Local tax rate \tab Fraction of all income going to local taxes. \cr
   Local government spending \tab per capita. \cr
   Progressivity \tab Measure of how much state income tax rates increase with income. \cr
   EITC \tab Measure of how much the state contributed to the Earned Income Tax Credit (a sort of negative income tax for very low-paid wage earners). \cr
   School expenditures \tab Average spending per pupil in public schools. \cr
   Student/teacher ratio \tab Number of students in public schools divided by number of teachers. \cr
   Test scores \tab Residuals from a linear regression of mean math and English test scores on household income per capita. \cr
   High school dropout rate \tab Residuals from a linear regression of the dropout rate on per-capita income. \cr
   Colleges per capita \tab  \cr
   College tuition \tab in-state, for full-time students \cr
   College graduation rate \tab Again, residuals from a linear regression of the actual graduation rate on household income per capita. \cr
   Labor force participation \tab Fraction of adults in the workforce. \cr
   Manufacturing \tab Fraction of workers in manufacturing. \cr
   Chinese imports \tab Growth rate in imports from China per worker between 1990 and 2000. \cr
   Teenage labor \tab fraction of those age 14–16 who were in the labor force. \cr
   Migration in \tab Migration into the community from elsewhere, as a fraction of 2000 population. \cr
   Migration out \tab Ditto for migration into other communities. \cr
   Foreign \tab fraction of residents born outside the US. \cr
   Social capital \tab Index combining voter turnout, participation in the census, and participation in community organizations. \cr
   Religious \tab Share of the population claiming to belong to an organized religious body. \cr
   Violent crime \tab Arrests per person per year for violent crimes. \cr
   Single motherhood \tab Number of single female households with children divided by the total number of households with children. \cr
   Divorced \tab Fraction of adults who are divorced. \cr
   Married \tab Ditto. \cr
   Longitude \tab Geographic coordinate for the center of the community \cr
   ID \tab A numerical code, identifying the community. \cr
   Name \tab the name of principal city or town. \cr
   State \tab the state of the principal city or town of the community. \cr
}
}
\keyword{datasets}
