% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-functions.R
\name{run_lab}
\alias{run_lab}
\title{Run a lab in your browser}
\usage{
run_lab(name = NULL)
}
\arguments{
\item{name}{the name of the lab to run. If omitted, lists all available labs.}
}
\value{

}
\description{
Run a lab in your browser
}
\details{
The intention is to make it easier to run a lab.
This is really just a wrapper for \code{learnr::run_tutorial(name, "UBCstat406labs")}
}
\examples{
run_lab() # Lists all our labs
\dontrun{
run_lab("rsquared") # runs the first lab
}
}
