% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{s301gradedist}
\alias{s301gradedist}
\title{S301 Grade Distributions}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 71 rows and 34 columns.
}
\usage{
s301gradedist
}
\description{
A dataset containing the grade distributions for each section of Indiana
University's
STAT-S301 (Business statistics) between 2015-2019.
}
\details{
\tabular{ll}{
   Variable \tab Description \cr
   term \tab the semester and year of the course \cr
   term_id \tab numeric code for the semester and year of the course \cr
   instructor \tab a unique number corresponding to the course instructor \cr
   n_students \tab number of students in the course \cr
   avg_grade \tab average final grade \cr
   avg_student_gpa \tab average gpa for enrolled students \cr
}


The remaining columns contain the percentage of students who earned A-D or other,
as well as the count of students earning each of the allowed letter grades.
}
\keyword{datasets}
