% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-functions.R
\name{show_labs}
\alias{show_labs}
\title{List all available labs (tutorials)}
\usage{
show_labs(package = "UBCstat406labs")
}
\arguments{
\item{package}{Default is "UBCstat406labs"}
}
\value{

}
\description{
This function lists all the labs available for Stat 406.
}
\details{
The intention is to make it easier to list all the labs we have.
This is really just a wrapper for \code{learnr::available_tutorials("UBCstat406labs")}
}
\examples{
show_labs()
}
