% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBCstat406labs-package.R
\docType{package}
\name{UBCstat406labs-package}
\alias{UBCstat406labs}
\alias{UBCstat406labs-package}
\title{UBCstat406labs: Tutorials and labs for UBC Stat 406 in the 2020-2021 online year}
\description{
This package is intended to be a one-stop install for students. The
    package contains all interactive labs for the semester. It also forces students
    to download/install various dependencies all at once, ensuring full compliance.
}
\author{
\strong{Maintainer}: Daniel McDonald \email{daniel@stat.ubc.ca}

}
\keyword{internal}
